requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
lmdb>=1.4.1
six>=1.16.0
matplotlib>=3.7.0
tqdm>=4.65.0
scikit-learn>=1.3.0

